+++ era5cli/fetch.py
@@ -78,15 +78,11 @@ class Fetch:
             or make the request to start downloading the data.
             `dryrun = True` will print the request to stdout. By default,
             the data will be downloaded.
-        prelimbe: bool
-            Whether to download the preliminary back extension (1950-1978).
-            Note that in this case, `years` must be between 1950 and
-            1978. `prelimbe = True` is incompatible with `land = True`.
         land: bool
             Whether to download data from the ERA5-Land dataset.
             Note that the ERA5-Land dataset starts in 1981.
             `land = True` is incompatible with the use of
-            `prelimbe = True` and `ensemble = True`.
+            `ensemble = True`.
         overwrite: bool
             Whether to overwrite existing files or not.
             Setting `overwrite = True` will make
@@ -117,7 +113,6 @@ class Fetch:
         splitmonths=False,
         merge=False,
         threads=None,
-        prelimbe=False,
         land=False,
         overwrite=False,
         dashed_vars=False,
@@ -173,9 +168,6 @@ class Fetch:
         """bool: Whether to get monthly averaged by hour of day
         (synoptic=True) or monthly means of daily means
         (synoptic=False)."""
-        self.prelimbe = prelimbe
-        """bool: Whether to select from the ERA5 preliminary back
-        extension which supports years from 1950 to 1978"""
         self.land = land
         """bool: Whether to download from the ERA5-Land
         dataset."""
@@ -186,24 +178,12 @@ class Fetch:
         files, or the normal names."""
 
         if self.merge and self.splitmonths:
-            raise ValueError(
-                "\nThe commands '--merge' and '--splitmonths' are not compatible with"
-                "\neach other. Please pick one of the two."
-            )
-
-        if self.prelimbe:
-            logging.warning(
-                "\n  The years of the ERA5 preliminary back extension (1950 - 1978) are"
-                "\n  now included in the main ERA5 products. The `--prelimbe` argument"
-                "\n  will be deprecated in a future release."
-                "\n  Please update your workflow accordingly."
-            )
+            self.splitmonths = False
 
         vars = list(self.variables)  # Use list() to avoid copying by reference
         if "geopotential" in vars and pressurelevels == ["surface"]:
             vars.remove("geopotential")
         if any([var in ref.PLVARS for var in vars]):
-            print(pressurelevels)
             self._check_levels()
 
         if self.period == "hourly" and request_too_large(self):
@@ -371,18 +351,6 @@ class Fetch:
         if self.synoptic:
             producttype += "_by_hour_of_day"
 
-        if not self.prelimbe:
-            return producttype
-
-        # Prelimbe has deviating product types for monthly data
-        if self.ensemble:
-            producttype = "members-"
-        else:
-            producttype = "reanalysis-"
-        if self.synoptic:
-            producttype += "synoptic-monthly-means"
-        else:
-            producttype += "monthly-means-of-daily-means"
         return producttype
 
     def _check_levels(self):
@@ -470,14 +438,6 @@ class Fetch:
 
         if self.land:
             name += "-land"
-        elif variable == "orography":
-            variable = "geopotential"
-            name += "-single-levels"
-            logging.warning(
-                "\n  The variable 'orography' has been deprecated by CDS. Use"
-                "\n  `--variables geopotential --levels surface` going forward."
-                "\n  The current query has been changed accordingly."
-            )
         elif self.pressure_levels == ["surface"]:
             name += "-single-levels"
         elif variable in ref.PLVARS:
@@ -490,13 +450,6 @@ class Fetch:
         if self.period == "monthly":
             name += "-monthly-means"
 
-        if self.prelimbe:
-            if self.land:
-                raise ValueError(
-                    "Back extension not available for ERA5-Land. "
-                    "ERA5-Land data is available from 1950 on."
-                )
-            name += "-preliminary-back-extension"
         return name, variable
 
     def _build_request(self, variable, years, months=None):