+++ tests/test_integration.py
@@ -18,34 +18,11 @@ def my_thing_mock():
 # combine calls with result and possible warning message
 call_result = [
     {
-        # orography is translated to geopotential in the query
-        "call": dedent(
-            """\
-            era5cli hourly --variables orography --startyear 2008 --dryrun
-            """
-        ),
-        "result": dedent(
-            """\
-            reanalysis-era5-single-levels {'variable': 'geopotential', 'year':
-            2008, 'month': ['01', '02', '03', '04', '05', '06', '07', '08',
-            '09', '10', '11', '12'], 'time': ['00:00', '01:00', '02:00',
-            '03:00', '04:00', '05:00', '06:00', '07:00', '08:00', '09:00',
-            '10:00', '11:00', '12:00', '13:00', '14:00', '15:00', '16:00',
-            '17:00', '18:00', '19:00', '20:00', '21:00', '22:00', '23:00'],
-            'format': 'netcdf', 'product_type': 'reanalysis', 'day': ['01',
-            '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12',
-            '13', '14', '15', '16', '17', '18', '19', '20', '21', '22', '23',
-            '24', '25', '26', '27', '28', '29', '30', '31']}
-            era5_orography_2008_hourly.nc"""
-        ),
-        "warn": "The variable 'orography' has been deprecated by CDS.",
-    },
-    {
         # geopotential needs '--levels surface' to be correctly interpreted
         "call": dedent(
             """\
             era5cli hourly --variables geopotential --startyear 2008 --dryrun
-            --levels surface"""
+             --splitmonths False --levels surface"""
         ),
         "result": dedent(
             """\
@@ -74,7 +51,7 @@ call_result = [
         "result": dedent(
             """\
             reanalysis-era5-pressure-levels {'variable': 'geopotential',
-            'year': 2008, 'month': ['01'], 'time': ['00:00', '01:00', '02:00',
+            'year': 2008, 'month': '01', 'time': ['00:00', '01:00', '02:00',
             '03:00', '04:00', '05:00', '06:00', '07:00', '08:00', '09:00',
             '10:00', '11:00', '12:00', '13:00', '14:00', '15:00', '16:00',
             '17:00', '18:00', '19:00', '20:00', '21:00', '22:00', '23:00'],
@@ -85,30 +62,9 @@ call_result = [
             '04', '05', '06', '07', '08', '09', '10', '11', '12', '13', '14',
             '15', '16', '17', '18', '19', '20', '21', '22', '23', '24', '25',
             '26', '27', '28', '29', '30', '31']}
-            era5_geopotential_2008_hourly.nc"""
+            era5_geopotential_2008-01_hourly.nc"""
         ),
         "warn": "Getting variable from pressure level data.",
-    },
-    {
-        # preliminary-back-extension is combined with monthly-means
-        "call": dedent(
-            """\
-            era5cli monthly --variables temperature --startyear 1960 --prelimbe
-            --dryrun"""
-        ),
-        "result": dedent(
-            """\
-            reanalysis-era5-pressure-levels-monthly-means-preliminary-back-extension
-            {'variable': 'temperature', 'year': 1960, 'month': ['01', '02',
-            '03', '04', '05', '06', '07', '08', '09', '10', '11', '12'],
-            'time': ['00:00'], 'format': 'netcdf', 'pressure_level': [1, 2, 3,
-            5, 7, 10, 20, 30, 50, 70, 100, 125, 150, 175, 200, 225, 250, 300,
-            350, 400, 450, 500, 550, 600, 650, 700, 750, 775, 800, 825, 850,
-            875, 900, 925, 950, 975, 1000], 'product_type':
-            'reanalysis-monthly-means-of-daily-means'}
-            era5_temperature_1960_monthly.nc"""
-        ),
-        "warn": "The years of the ERA5 preliminary back extension",
     },
     {
         # era5-Land is combined with monthly means
@@ -148,7 +104,7 @@ def test_main(call_result, capsys, caplog):
     with caplog.at_level(logging.INFO):
         with mock.patch(
             "era5cli.fetch.key_management.load_era5cli_config",
-            return_value=("url", "key:uid"),
+            return_value=("url", "key"),
         ):
             main(call)
     captured = capsys.readouterr().out